This was a weird little bug. Angulars docs say that when using $http() you should hook into .success and .error, which is fine and works. BUT it also has .then and .catch methods for inter-compatibility in a promise chain. Sadly, if you use BOTH success AND then, both will get called instead of chaining (and they get handed different inputs!)
So; in this code, a $http call is fired which starts up a promise chain off of its success method. This first promise is then returned whilst request is in progress. Upon return, a then is attached which DOESN'T append to the chain as expected, and instead starts a new chain all of its own, over-writing the .success based chain. As a result, when a server connection exists the controller gets an http response passed in and works fine. When an error occurred it was getting a pouchDb response and then reading it wrong as it was setup for an http response.

The fix was therefore three stages -
    One; re-hook the http onto .then() instead of .success() to get this chain running.
    Two: Fix the code inside the new .then() block to handle the full http response instead of the cleaned up data response which .success() gets and have it return from pouchDB
    Three: Fix the controller to expect a pouchDb response every time.

Working code!